pyproject.toml
[project]
name = "coh_npc_voices"
version = "1.1.0"
authors = [
  { name="Jason Kane" },
]
description = "A simple python program that lets you hear City of Heroes NPCs talk"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Environment :: Win32 (MS Windows)",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows"
]
dependencies = [
    "pypiwin32",
    "tts @ git+https://github.com/DeepHorizons/tts@master#11cabe8",
    "voicebox-tts[google-cloud-tts]",
    "google-cloud-texttospeech",
    "voicebox-tts",
    "wheel",
    "setuptools",
    "matplotlib",
    "sqlalchemy",
    "sqlalchemy-utils",
    "alembic"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.urls]
Homepage = "https://github.com/jason-kane/coh_npc_voices"
Issues = "https://github.com/jason-kane/coh_npc_voices/issues"

[project.scripts]
coh_voices = "coh_npc_voices.voice_editor:main"
sidekick = "coh_npc_voices.sidekick:main"

[tool.hatch.metadata]
allow-direct-references = true